[Security Solution][Alert details] - move threat intelligence to flyout folder
