--- x-pack/plugins/security_solution/public/common/components/event_details/cti_details/threat_summary_table.tsx
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import type { AnyStyledComponent } from 'styled-components';
-import styled from 'styled-components';
-import { EuiInMemoryTable } from '@elastic/eui';
-
-export const ThreatSummaryTable = styled(EuiInMemoryTable as unknown as AnyStyledComponent)`
-  .euiTableHeaderCell,
-  .euiTableRowCell {
-    border: none;
-  }
-  .euiTableHeaderCell .euiTableCellContent {
-    padding: 0;
-  }
-`;
