+++ x-pack/plugins/security_solution/public/flyout/document_details/left/hooks/use_threat_intelligence_details.ts
@@ -6,6 +6,7 @@
  */
 
 import { useMemo } from 'react';
+import { SecurityPageName } from '@kbn/deeplinks-security';
 import type { RunTimeMappings } from '../../../../../common/api/search_strategy';
 import type { CtiEnrichment, EventFields } from '../../../../../common/search_strategy';
 import { useBasicDataFromDetailsData } from '../../../../timelines/components/side_panel/event_details/helpers';
@@ -14,10 +15,8 @@ import {
   getEnrichmentFields,
   parseExistingEnrichments,
   timelineDataToEnrichment,
-} from '../../../../common/components/event_details/cti_details/helpers';
-import { SecurityPageName } from '../../../../../common/constants';
+} from '../../shared/utils/threat_intelligence';
 import { SourcererScopeName } from '../../../../sourcerer/store/model';
-
 import { useInvestigationTimeEnrichment } from '../../../../common/containers/cti/event_enrichment';
 import { useTimelineEventsDetails } from '../../../../timelines/containers/details';
 import { useSourcererDataView } from '../../../../sourcerer/containers';
@@ -25,15 +24,42 @@ import { useRouteSpy } from '../../../../common/utils/route/use_route_spy';
 import { useDocumentDetailsContext } from '../../shared/context';
 
 export interface ThreatIntelligenceDetailsValue {
+  /**
+   *
+   */
   enrichments: CtiEnrichment[];
+  /**
+   *
+   */
   eventFields: EventFields;
+  /**
+   *
+   */
   isEnrichmentsLoading: boolean;
+  /**
+   *
+   */
   isEventDataLoading: boolean;
+  /**
+   *
+   */
   isLoading: boolean;
+  /**
+   *
+   */
   range: {
+    /**
+     *
+     */
     from: string;
+    /**
+     *
+     */
     to: string;
   };
+  /**
+   *
+   */
   setRange: (range: { from: string; to: string }) => void;
 }
