+++ x-pack/test/security_solution_cypress/cypress/screens/alerts_details.ts
@@ -5,11 +5,8 @@
  * 2.0.
  */
 
-export const ALERT_FLYOUT = '[data-test-subj="timeline:details-panel:flyout"]';
-
-export const CELL_TEXT = '.euiText';
-
-export const ENRICHMENT_QUERY_RANGE_PICKER = '[data-test-subj="enrichment-query-range-picker"]';
+export const ENRICHMENT_QUERY_RANGE_PICKER =
+  '[data-test-subj="securitySolutionFlyoutThreatIntelligenceEnrichmentRangePicker"]';
 
 export const ENRICHMENT_QUERY_START_INPUT = '.start-picker';
 
@@ -17,23 +14,21 @@ export const ENRICHMENT_QUERY_END_INPUT = '.end-picker';
 
 export const FILTER_INPUT = '[data-test-subj="eventDetails"] .euiFieldSearch';
 
-export const INDICATOR_MATCH_ENRICHMENT_SECTION = '[data-test-subj="threat-match-detected"]';
+export const INDICATOR_MATCH_ENRICHMENT_SECTION =
+  '[data-test-subj="securitySolutionFlyoutThreatIntelligenceThreatMatchDetected"]';
 
 export const INVESTIGATION_TIME_ENRICHMENT_SECTION =
-  '[data-test-subj="enriched-with-threat-intel"]';
+  '[data-test-subj="securitySolutionFlyoutThreatIntelligenceEnrichedWithThreatIntel"]';
+
+export const ENRICHMENT_SECTION_BUTTON_CONTENT =
+  '[data-test-subj="securitySolutionFlyoutThreatIntelligenceEnrichmentButtonContent"]';
 
 export const JSON_VIEW_TAB = '[data-test-subj="jsonViewTab"]';
-
-export const JSON_TEXT = '[data-test-subj="kibanaCodeEditor"]';
-
-export const OVERVIEW_RULE = '[data-test-subj="eventDetails"] [data-test-subj="ruleName"]';
 
 export const OVERVIEW_STATUS = '[data-test-subj="eventDetails"] [data-test-subj="alertStatus"]';
 
 export const EVENT_DETAILS_ALERT_STATUS_POPOVER =
   '[data-test-subj="event-details-alertStatusPopover"]';
-
-export const SUMMARY_VIEW = '[data-test-subj="summary-view"]';
 
 export const TABLE_CELL = '.euiTableRowCell';
 
@@ -41,18 +36,12 @@ export const CELL_EXPAND_VALUE = '[data-test-subj="euiDataGridCellExpandButton"]
 
 export const TABLE_TAB = '[data-test-subj="tableTab"]';
 
-export const TABLE_CONTAINER = '[data-test-subj="event-fields-browser"]';
-
 export const TABLE_ROWS = '.euiTableRow';
 
-export const THREAT_DETAILS_ACCORDION = '.euiAccordion__triggerWrapper';
+export const THREAT_DETAILS_VIEW =
+  '[data-test-subj="securitySolutionFlyoutThreatIntelligenceEnrichmentAccordionTable-0"]';
 
-export const THREAT_DETAILS_VIEW = '[data-test-subj="threat-details-view-0"]';
-
-export const THREAT_INTEL_TAB = '[data-test-subj="threatIntelTab"]';
-
-export const UPDATE_ENRICHMENT_RANGE_BUTTON = '[data-test-subj="enrichment-button"]';
+export const UPDATE_ENRICHMENT_RANGE_BUTTON =
+  '[data-test-subj="securitySolutionFlyoutThreatIntelligenceEnrichmentRefreshButton"]';
 
 export const ENRICHED_DATA_ROW = `[data-test-subj='EnrichedDataRow']`;
-
-export const COPY_ALERT_FLYOUT_LINK = `[data-test-subj="copy-alert-flyout-link"]`;
